Update README.md
This is program is used for detecting edges in python. It accepts the video using the cv2 module from open cv, grays it and blurs the image for easy visibility of edges, using the same. this project makes it easier for artists to notice the subtle lines that can be missed by the human eye and would help in creating a more aesthetic filters for photos.